backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.10.1
betterproto==2.0.0b4
asyncpg==0.29.0
alembic==1.13.1
openai==1.3.5
httpx==0.25.2
python-dotenv==1.0.0
uuid==1.30
typing-extensions==4.8.0
psycopg2-binary==2.9.9
pytest==7.4.3
